+++ src/mjlab/tasks/tracking/tracking_env_cfg.py
@@ -48,6 +48,10 @@ def make_tracking_env_cfg() -> ManagerBasedRlEnvCfg:
     "command": ObservationTermCfg(
       func=mdp.generated_commands, params={"command_name": "motion"}
     ),
+    "motion_lookahead": ObservationTermCfg(
+      func=mdp.motion_lookahead,
+      params={"command_name": "motion"},
+    ),
     "motion_anchor_pos_b": ObservationTermCfg(
       func=mdp.motion_anchor_pos_b,
       params={"command_name": "motion"},